Add trusted extension component
